requirements.txt
adafruit-io==2.5.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cryptography==3.4.7
Deprecated==1.2.12
firebase==3.0.1
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
jwcrypto==0.9.1
Naked==0.1.31
oauth2client==4.1.3
paho-mqtt==1.5.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
python-jwt==3.3.0
PyYAML==5.4.1
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
shellescape==3.8.1
six==1.16.0
sseclient==0.0.27
urllib3==1.26.6
wrapt==1.12.1
